Update build_vocab.py
No `max_size` limit for vocabulary selection from BPE-segmented data. 
For BPE, the number of merges is fixed (here: uses `max_size`, we should probably rename it for BPE to make clear that it's the number of merges and not the final vocabulary size), and the resulting vocabulary size can be slightly larger, so it's safer to include all subwords of the training set in the vocabulary.